Edit code to look simplified
Got rid of all the old JS and wrote it in a more compact format.

Most of the JS is written in an array so it's easier to call
certain items and functions and to add more buildings in the future.

Removed the HTML table as everything is still able to work the same
without it.